# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM APP v2 API",
        "contact": {
            "name": "Dev CRM Team",
            "email": "longnguyen@gearment.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/sample/{category}/things": {
            "get": {
                "tags": [
                    "yourtag"
                ],
                "operationId": "/sample/category/things",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "The category parameter in path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "criteria",
                        "in": "query",
                        "description": "Some optional other parameter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns some sample category things",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/foo": {
            "get": {
                "tags": [
                    "yourtag"
                ],
                "operationId": "/foo",
                "responses": {
                    "200": {
                        "description": "Returns some sample category things",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/todos": {
            "get": {
                "tags": [
                    "yourtag"
                ],
                "operationId": "/todos",
                "responses": {
                    "200": {
                        "description": "Returns some sample todos things",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        }
    }
}